requirements.txt
pycrypto
fakeuseragent